Disable libavserver (Closes: #734335)
It turns out that avserver is pretty much unmaintained these days and
contains obvious bugs. The next Libav upstream release has it already
released, and further maintenance is currently rather unlikely. With
this in mind, the libav package proactively disables the avserver build
anticipating its removal upstream.

If you are interested in having avserver back and functional, please
help upstream with reinstantiating it.